refactor fetch transfers history
