+++ src/pages/network/tron/Token.jsx
@@ -1,4 +1,4 @@
-import { useContext, useEffect, useMemo, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import { ethers } from "ethers";
 import styled from "styled-components";
 
@@ -155,21 +155,15 @@ const Token = () => {
   const [bandwidthEstimation, setBandwidthEstimation] = useState(0);
   const [energyEstimation, setEnergyEstimation] = useState(0);
 
-  const [history, setHistory] = useState([]);
-
   const {
     account,
     selectedTokenAddress,
     selectToken,
-    tokenBalances,
+    token,
     fetchBalances,
+    transfers,
     loading,
   } = useContext(StoreContext);
-
-  const token = useMemo(
-    () => tokenBalances.find((t) => t.address === selectedTokenAddress),
-    [tokenBalances, selectedTokenAddress]
-  );
 
   const { debounce } = useDebounce();
   const estimateGas = debounce(async () => {
@@ -222,14 +216,6 @@ const Token = () => {
       setHasErrorInput(true);
     }
   }, 500);
-
-  useEffect(() => {
-    if (!token.native) {
-      account.queryTokenHistory(selectedTokenAddress).then((result) => {
-        setHistory(result || []);
-      });
-    }
-  }, []);
 
   useEffect(() => {
     estimateGas();
@@ -387,8 +373,8 @@ const Token = () => {
         </Submit>
       </Box>
 
-      {history.length > 0 && (
-        <ActivityTable account={account} history={history} />
+      {transfers.length > 0 && (
+        <ActivityTable account={account} history={transfers} />
       )}
     </Container>
   );
